@@ -4,16 +4,15 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Tests
+namespace Tests;
+
+/// <summary>
+/// the interface that supports simple Number plus and minus test
+/// </summary>
+public interface INumberStory
 {
     /// <summary>
-    /// the interface that supports simple Number plus and minus test
+    /// the Number being modified
     /// </summary>
-    public interface INumberStory
-    {
-        /// <summary>
-        /// the Number being modified
-        /// </summary>
-        public int Number { get; set; }
-    }
+    public int Number { get; set; }
 }
