@@ -5,32 +5,31 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Tests
+namespace Tests;
+
+/// <summary>
+/// the number minus event
+/// </summary>
+public class MinusEvent : IEvent
 {
     /// <summary>
-    /// the number minus event
+    /// the testing story
     /// </summary>
-    public class MinusEvent : IEvent
-    {
-        /// <summary>
-        /// the testing story
-        /// </summary>
-        private INumberStory _story;
+    private INumberStory _story;
 
-        /// <summary>
-        /// the constructor
-        /// </summary>
-        /// <param name="story">the testing story</param>
-        public MinusEvent(INumberStory story) { _story = story; }
+    /// <summary>
+    /// the constructor
+    /// </summary>
+    /// <param name="story">the testing story</param>
+    public MinusEvent(INumberStory story) { _story = story; }
 
-        /// <summary>
-        /// the event time in the simulation
-        /// </summary>
-        public TimeSpan EventTime { get; set; }
+    /// <summary>
+    /// the event time in the simulation
+    /// </summary>
+    public TimeSpan EventTime { get; set; }
 
-        /// <summary>
-        /// process the number minus operation
-        /// </summary>
-        public void Process() { _story.Number--; }
-    }
+    /// <summary>
+    /// process the number minus operation
+    /// </summary>
+    public void Process() { _story.Number--; }
 }
